+++ src/Type/MessageId.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Type;
+
+use Vjik\TelegramBot\Api\ParseResult\ValueHelper;
+
+/**
+ * https://core.telegram.org/bots/api#messageid
+ */
+final readonly class MessageId
+{
+    public function __construct(
+        public int $messageId,
+    ) {
+    }
+
+    public static function fromTelegramResult(mixed $result): self
+    {
+        ValueHelper::assertArrayResult($result);
+        return new self(
+            ValueHelper::getInteger($result, 'message_id'),
+        );
+    }
+}